experimental feature: hotreload by pressing F7.
Depending on the package and the current menu's state, it might break, crash, stop accepting input, etc. Some of these cases are mitigated by not hot-reloading packages that don't have an associated menu screen.
These packages are very likely to be referenced by data that is not owned by the package data, so we can not reinitialize it, leading to dead pointers.